pyproject.toml
[project]
name = "liangjian"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.8"
dependencies = [
    "grpcio==1.28.1",
    "numpy>=1.24.4",
    "pip>=25.0.1",
    "protobuf==3.11.3",
    "pyinquirer==1.0.3",
    "websockets==8.1",
    "xmltodict~=0.12.0",
]
